float instead of int for cjfee
A poorly coded or malicious maker bot script may be able to cause all potential takers to hang when seeking available orders, with something like

    ValueError: invalid literal for int() with base 10: '0.1337'

This is done by making certain changes to how the maker order is announced to potential takers in IRC.

Change int to float instead possibly to avoid hanging?  Not sure what else this may affect or if this is the best solution, in any case, but this is a potential way to deny service to takers, so please review ASAP.

+++ lib/common.py
@@ -335,7 +335,7 @@ class BitcoinCoreWallet(AbstractWallet):
 def calc_cj_fee(ordertype, cjfee, cj_amount):
 	real_cjfee = None
 	if ordertype == 'absorder':
-		real_cjfee = int(cjfee)
+		real_cjfee = float(cjfee)
 	elif ordertype == 'relorder':
 		real_cjfee = int((Decimal(cjfee) * Decimal(cj_amount)).quantize(Decimal(1)))
 	else:
